pyproject.toml
[project]
name = "streamlit-langgraph"
version = "0.1.0"
description = "Steamlit project using langgraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio==3.4.3",
    "langchain==0.3.18",
    "langchain-openai==0.3.15",
    "langgraph==0.3.1",
    "langsmith==0.3.38",
    "python-dotenv==1.1.0",
    "streamlit==1.45.0",
]
